requirements.txt
google-generativeai>=0.8.0
feedparser>=6.0.12
openpyxl>=3.1.5
python-dotenv>=1.2.2
